last

+++ carsAPI/Controllers/CarsController.cs
@@ -3,15 +3,21 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Threading;
 using System.Web.Http;
 using carsAPI.Models;
 using Newtonsoft.Json;
+using carsAPI.Controllers;
+
 
 namespace carsAPI.Controllers
 {
+    [BasicAuthentication]
     [RoutePrefix("api/cars")]
     public class CarsController : ApiController
     {
+        public static string ThreadUser = Thread.CurrentPrincipal.Identity.Name;
+        
         // GET: api/cars
         [HttpGet]
         [Route("find")]
@@ -27,11 +33,12 @@ namespace carsAPI.Controllers
 
                         carNumber = p.carNumber,
                         carType = p.carType ?? int.MinValue,
-                        imagePath = p.imagePath,
                         isAvailable = p.isAvailable,
                         isUndamaged = p.isUndamaged,
                         region = p.region,
-                        mileage = p.mileage ?? int.MinValue
+                        mileage = p.mileage ?? int.MinValue,
+                        image = p.image
+                        
 
 
                     }).ToList();
@@ -60,31 +67,39 @@ namespace carsAPI.Controllers
         [Route("create")]
         public HttpResponseMessage create(carEntity carEntity)
         {
-            using ( var db = new rentcarsEntities() )
+            using (var db = new rentcarsEntities())
             {
-                try
-                {  
-                    var response = new HttpResponseMessage(HttpStatusCode.OK);
-                    var car = new car()
-                    {
-                        carNumber = carEntity.carNumber,
-                        carType = carEntity.carType,
-                        imagePath = carEntity.imagePath,
-                        isAvailable = carEntity.isAvailable,
-                        isUndamaged = carEntity.isUndamaged,
-                        region = carEntity.region,
-                        mileage = carEntity.mileage
-                    };
-
-                    db.cars.Add(car);
-                    db.SaveChanges();
-                    return response;
-
-                }
-                catch
+                if ( ThreadUser == "admin")
                 {
+                    try
+                    {
+                        var response = new HttpResponseMessage(HttpStatusCode.OK);
+                        var car = new car()
+                        {
+                            carNumber = carEntity.carNumber,
+                            carType = carEntity.carType,
+                            isAvailable = carEntity.isAvailable,
+                            isUndamaged = carEntity.isUndamaged,
+                            region = carEntity.region,
+                            mileage = carEntity.mileage
 
+                        };
+
+                        db.cars.Add(car);
+                        db.SaveChanges();
+                        return response;
+
+                    }
+                    catch
+                    {
+
+                        return new HttpResponseMessage(HttpStatusCode.BadRequest);
+                    }
+                }
+                else
+                {
                     return new HttpResponseMessage(HttpStatusCode.BadRequest);
+
                 }
             }
         }
@@ -96,23 +111,29 @@ namespace carsAPI.Controllers
         {
             using (var db = new rentcarsEntities())
             {
-                try
+                if (ThreadUser == "admin")
                 {
-                    var response = new HttpResponseMessage(HttpStatusCode.OK);
-                    var currentCar = db.cars.SingleOrDefault(p => p.carNumber == car.carNumber);
-                    currentCar.carNumber = car.carNumber;
-                    currentCar.carType = car.carType;
-                    currentCar.imagePath = car.imagePath;
-                    currentCar.isAvailable = car.isAvailable;
-                    currentCar.isUndamaged = car.isUndamaged;
-                    currentCar.region = car.region;
-                    currentCar.mileage = car.mileage;                   
-                    db.SaveChanges();
-                    return response;
-                }
-                catch 
-                {
+                    try
+                    {
+                        var response = new HttpResponseMessage(HttpStatusCode.OK);
+                        var currentCar = db.cars.SingleOrDefault(p => p.carNumber == car.carNumber);
+                        currentCar.carNumber = car.carNumber;
+                        currentCar.carType = car.carType;
+                        currentCar.isAvailable = car.isAvailable;
+                        currentCar.isUndamaged = car.isUndamaged;
+                        currentCar.region = car.region;
+                        currentCar.mileage = car.mileage;
+                        db.SaveChanges();
+                        return response;
+                    }
+                    catch
+                    {
 
+                        return new HttpResponseMessage(HttpStatusCode.BadRequest);
+                    }
+                }
+                else
+                {
                     return new HttpResponseMessage(HttpStatusCode.BadRequest);
                 }
             }
@@ -123,26 +144,35 @@ namespace carsAPI.Controllers
         [Route("delete/{carNumber}")]
         public HttpResponseMessage Delete(string carNumber)
         {
-        using (var db = new rentcarsEntities())
-        {
-
-            try
+            using (var db = new rentcarsEntities())
             {
-               
-                var response = new HttpResponseMessage(HttpStatusCode.OK);
+                if (ThreadUser == "admin")
+                {
 
-                    var car = db.cars.SingleOrDefault(p => p.carNumber == carNumber);
-                    db.cars.Remove(car);
-                    db.SaveChanges();
-                    return response;
+                    try
+                    {
 
-            }
-            catch
-            {
+                        var response = new HttpResponseMessage(HttpStatusCode.OK);
 
-                return new HttpResponseMessage(HttpStatusCode.BadRequest);
+                        var car = db.cars.SingleOrDefault(p => p.carNumber == carNumber);
+                        db.cars.Remove(car);
+                        db.SaveChanges();
+                        return response;
+
+                    }
+                    catch
+                    {
+
+                        return new HttpResponseMessage(HttpStatusCode.BadRequest);
+                    }
+                }
+
+                else
+                {
+                    return new HttpResponseMessage(HttpStatusCode.BadRequest);
+                }
             }
         }
-    }
+
     }
 }
